pyproject.toml
[project]
name = "openai-agents-streaming-api"
version = "0.1.0"
description = "A FastAPI backend for serving AI agents."
authors = [{name = "Muhammad Ahmad Shoukat", email = "ahmadshaukat_4@outlook.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
    "ty>=0.0.1a26",
    "pytest>=8.0.0",
]
